Move summary page to route /registration/<id>/signup-group/completed
