+++ src/domain/signupGroup/SignupGroupCompletedPage.tsx
@@ -7,21 +7,21 @@ import LoadingSpinner from '../../common/components/loadingSpinner/LoadingSpinne
 import SuccessTemplate from '../../common/components/successTemplate/SuccessTemplate';
 import useLocale from '../../hooks/useLocale';
 import MainContent from '../app/layout/mainContent/MainContent';
+import ConfirmationMessage from '../enrolment/confirmationMessage/ConfirmationMessage';
+import { ENROLMENT_QUERY_PARAMS } from '../enrolment/constants';
+import useEventAndRegistrationData from '../enrolment/hooks/useEventAndRegistrationData';
 import { Event } from '../event/types';
 import { getEventFields } from '../event/utils';
 import NotFound from '../notFound/NotFound';
 import { Registration } from '../registration/types';
 import { getRegistrationFields } from '../registration/utils';
-import ConfirmationMessage from './confirmationMessage/ConfirmationMessage';
-import { ENROLMENT_QUERY_PARAMS } from './constants';
-import useEventAndRegistrationData from './hooks/useEventAndRegistrationData';
 
 type Props = {
   event: Event;
   registration: Registration;
 };
 
-const EnrolmentCompletedPage: React.FC<Props> = ({ event, registration }) => {
+const SignupGroupCompletedPage: React.FC<Props> = ({ event, registration }) => {
   const { query } = useRouter();
   const { [ENROLMENT_QUERY_PARAMS.REDIRECT_URL]: redirectUrl } = query;
   const { t } = useTranslation(['enrolment']);
@@ -66,13 +66,13 @@ const EnrolmentCompletedPage: React.FC<Props> = ({ event, registration }) => {
   );
 };
 
-const EnrolmentCompletedPageWrapper: React.FC = () => {
+const SignupGroupCompletedPageWrapper: React.FC = () => {
   const { event, isLoading, registration } = useEventAndRegistrationData();
 
   return (
     <LoadingSpinner isLoading={isLoading}>
       {event && registration ? (
-        <EnrolmentCompletedPage event={event} registration={registration} />
+        <SignupGroupCompletedPage event={event} registration={registration} />
       ) : (
         <NotFound />
       )}
@@ -80,4 +80,4 @@ const EnrolmentCompletedPageWrapper: React.FC = () => {
   );
 };
 
-export default EnrolmentCompletedPageWrapper;
+export default SignupGroupCompletedPageWrapper;
